Add a clusterpolicy to set a default ttl for jobs that lack it
For starters we will set the ttl to three days as a max age any job can linger
after it has finished. Both successfully or with a failure.

If a Job includes a ttl that will be honored, this only sets a default that gets
injected at Job submission time if the Job lacks a definition.

This is to prevent or ameliorate jobs that might linger too long and use disk
space needlessly.