Comment out podSecurity for cray-sysmgmt-health-prometheus-node-exporter*
The cray-sysmgmt-health-prometheus-node-exporter Deployment has
an AppArmor securityContext.  This changed in k8s 1.30.x, and
kyverno 1.13.x does not handle the change.  Thus the only way
to get rid of the violation is to excempt the Deployment from
the entire PSS policy, and that is done by commenting out the
podSecurity section.  Once kyverno 1.14.x comes out with a fix,
then we can restore the podSecurity section.