Add policy exception for Job cray-sls-init-load-nameserver-getter
When doing a fresh install, this HostPath issue popped up in the
"kubectl get events -A" output.  It does not show up in the
policy reports, because the job is created, run, and removed
during the install, so the policy reports for it were also
removed when the pod/job was removed.